added script conv_0d.py
