+++ scripts/convolution/demos/conv_2d_demoy.py
