+++ scripts/convolution/demos/gauss_sampling_demo.py
